+++ common/src/main/java/io/github/flemmli97/runecraftory/common/effects/SleepEffect.java
@@ -1,7 +1,7 @@
 package io.github.flemmli97.runecraftory.common.effects;
 
 import io.github.flemmli97.runecraftory.common.entities.SleepingEntity;
-import io.github.flemmli97.runecraftory.common.network.S2CEntityDataSync;
+import io.github.flemmli97.runecraftory.mixinhelper.MobNoAIHandler;
 import io.github.flemmli97.runecraftory.platform.Platform;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
@@ -17,15 +17,20 @@ public class SleepEffect extends MobEffect {
         super(MobEffectCategory.HARMFUL, 0);
     }
 
-    private static void sendSleepPacket(LivingEntity entity, boolean flag) {
-        Platform.INSTANCE.sendToTrackingAndSelf(new S2CEntityDataSync(entity.getId(), S2CEntityDataSync.Type.SLEEP, flag), entity);
+    private static void applySleep(LivingEntity entity, boolean flag) {
+        Platform.INSTANCE.getEntityData(entity).ifPresent(data -> data.setSleeping(entity, flag));
     }
 
     @Override
     public void applyEffectTick(LivingEntity living, int amplifier) {
         if (!(living instanceof Player player) || !player.getAbilities().invulnerable) {
             if (!living.noPhysics)
-                living.setDeltaMovement(new Vec3(living.getDeltaMovement().x, -0.08, -living.getDeltaMovement().z));
+                living.setDeltaMovement(new Vec3(living.getDeltaMovement().x, -0.08, living.getDeltaMovement().z));
+            // Setting no ai prevents mobs moving/getting moved. So we do it here
+            if (living instanceof MobNoAIHandler mob) {
+                mob.setIgnoreNoAI();
+                living.travel(new Vec3(living.xxa, living.yya, living.zza));
+            }
         }
         MobEffectInstance eff = living.getEffect(this);
         if (eff.getDuration() > 200) {
@@ -43,8 +48,7 @@ public class SleepEffect extends MobEffect {
     public void removeAttributeModifiers(LivingEntity entity, AttributeMap attributeMap, int amplifier) {
         if (entity instanceof SleepingEntity sleeping)
             sleeping.setSleeping(false);
-        entity.setSilent(false);
-        sendSleepPacket(entity, false);
+        applySleep(entity, false);
         super.removeAttributeModifiers(entity, attributeMap, amplifier);
     }
 
@@ -52,8 +56,8 @@ public class SleepEffect extends MobEffect {
     public void addAttributeModifiers(LivingEntity entity, AttributeMap attributeMap, int amplifier) {
         if (entity instanceof SleepingEntity sleeping)
             sleeping.setSleeping(true);
-        entity.setSilent(true);
-        sendSleepPacket(entity, true);
+        entity.setDeltaMovement(Vec3.ZERO);
+        applySleep(entity, true);
         super.addAttributeModifiers(entity, attributeMap, amplifier);
     }
 }