+++ common/src/main/java/io/github/flemmli97/runecraftory/common/effects/TrueInvis.java
@@ -1,6 +1,5 @@
 package io.github.flemmli97.runecraftory.common.effects;
 
-import io.github.flemmli97.runecraftory.common.network.S2CEntityDataSync;
 import io.github.flemmli97.runecraftory.platform.ExtendedEffect;
 import io.github.flemmli97.runecraftory.platform.Platform;
 import net.minecraft.world.effect.MobEffect;
@@ -26,13 +25,13 @@ public class TrueInvis extends MobEffect implements ExtendedEffect {
 
     @Override
     public void removeAttributeModifiers(LivingEntity entity, AttributeMap manager, int amplifier) {
-        Platform.INSTANCE.sendToTrackingAndSelf(new S2CEntityDataSync(entity.getId(), S2CEntityDataSync.Type.INVIS, false), entity);
+        Platform.INSTANCE.getEntityData(entity).ifPresent(data -> data.setInvis(entity, false));
         super.removeAttributeModifiers(entity, manager, amplifier);
     }
 
     @Override
     public void addAttributeModifiers(LivingEntity entity, AttributeMap manager, int amplifier) {
-        Platform.INSTANCE.sendToTrackingAndSelf(new S2CEntityDataSync(entity.getId(), S2CEntityDataSync.Type.INVIS, true), entity);
+        Platform.INSTANCE.getEntityData(entity).ifPresent(data -> data.setInvis(entity, true));
         super.addAttributeModifiers(entity, manager, amplifier);
     }
 }