+++ common/src/main/java/io/github/flemmli97/runecraftory/mixin/EntityMixin.java
@@ -1,10 +1,12 @@
 package io.github.flemmli97.runecraftory.mixin;
 
+import io.github.flemmli97.runecraftory.common.attachment.EntityData;
 import io.github.flemmli97.runecraftory.common.entities.GateEntity;
 import io.github.flemmli97.runecraftory.mixinhelper.MixinUtils;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.MoverType;
+import net.minecraft.world.entity.Pose;
 import net.minecraft.world.phys.Vec3;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -30,7 +32,7 @@ public abstract class EntityMixin {
 
     @ModifyVariable(method = "move", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/entity/Entity;getMovementEmission()Lnet/minecraft/world/entity/Entity$MovementEmission;"), ordinal = 1)
     private Vec3 onStepLiving(Vec3 orig, MoverType type, Vec3 pos) {
-        if (((Object) this) instanceof LivingEntity living) {
+        if ((Object) this instanceof LivingEntity living) {
             this.runecraftoryMoveTracker += orig.lengthSqr();
             // Get the next step value independent of current move dist
             float pre = living.moveDist;
@@ -44,6 +46,13 @@ public abstract class EntityMixin {
             }
         }
         return orig;
+    }
+
+    @Inject(method = "getPose", at = @At("HEAD"), cancellable = true)
+    private void checkSleepingPose(CallbackInfoReturnable<Pose> info) {
+        if ((Object) this instanceof LivingEntity living && EntityData.getSleepState(living) == EntityData.SleepState.VANILLA) {
+            info.setReturnValue(Pose.SLEEPING);
+        }
     }
 
     @Shadow