+++ common/src/main/java/io/github/flemmli97/runecraftory/mixin/EntityRendererDispatcherMixin.java
@@ -11,7 +11,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(EntityRenderDispatcher.class)
-public class EntityRendererDispatcherMixin {
+public abstract class EntityRendererDispatcherMixin {
 
     @Inject(method = "shouldRender", at = @At("HEAD"), cancellable = true)
     private <E extends Entity> void onInvisCheck(E entity, Frustum frustum, double camX, double camY, double camZ, CallbackInfoReturnable<Boolean> info) {