+++ common/src/main/java/io/github/flemmli97/runecraftory/mixin/HoeItemMixin.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(HoeItem.class)
-public class HoeItemMixin {
+public abstract class HoeItemMixin {
 
     @Inject(method = "useOn", at = @At(value = "INVOKE", target = "Lnet/minecraft/world/item/context/UseOnContext;getItemInHand()Lnet/minecraft/world/item/ItemStack;"))
     private void onHoeUse(UseOnContext ctx, CallbackInfoReturnable<InteractionResult> info) {