+++ common/src/main/java/io/github/flemmli97/runecraftory/mixin/LivingEntityRendererMixin.java
@@ -3,12 +3,11 @@ package io.github.flemmli97.runecraftory.mixin;
 import com.mojang.blaze3d.vertex.PoseStack;
 import io.github.flemmli97.runecraftory.common.attachment.EntityData;
 import io.github.flemmli97.runecraftory.mixinhelper.ClientMixinUtils;
-import net.minecraft.client.renderer.MultiBufferSource;
 import net.minecraft.client.renderer.entity.LivingEntityRenderer;
 import net.minecraft.core.Direction;
 import net.minecraft.world.entity.LivingEntity;
-import net.minecraft.world.entity.Pose;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
@@ -17,17 +16,14 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 @Mixin(LivingEntityRenderer.class)
 public abstract class LivingEntityRendererMixin<T extends LivingEntity> {
 
-    @Inject(method = "render", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/renderer/entity/LivingEntityRenderer;getBob(Lnet/minecraft/world/entity/LivingEntity;F)F"))
-    public void render(T entity, float entityYaw, float partialTicks, PoseStack poseStack, MultiBufferSource buffer, int packedLight, CallbackInfo info) {
-        ClientMixinUtils.translateSleepingEntity(entity, poseStack, partialTicks);
+    @Inject(method = "setupRotations", at = @At(value = "RETURN"))
+    public void render(T entity, PoseStack poseStack, float ageInTicks, float rotationYaw, float partialTicks, CallbackInfo info) {
+        if (this.getFlipDegrees(entity) != 0)
+            ClientMixinUtils.translateSleepingEntity(entity, poseStack);
     }
 
-    @ModifyVariable(method = "setupRotations", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/entity/LivingEntity;getPose()Lnet/minecraft/world/entity/Pose;"))
-    private Pose pose(Pose original, T entityLiving) {
-        if (EntityData.getSleepState(entityLiving) == EntityData.SleepState.VANILLA)
-            return Pose.SLEEPING;
-        return original;
-    }
+    @Shadow
+    abstract float getFlipDegrees(T livingEntity);
 
     @ModifyVariable(method = "setupRotations", at = @At(value = "INVOKE_ASSIGN", target = "Lnet/minecraft/world/entity/LivingEntity;getBedOrientation()Lnet/minecraft/core/Direction;"))
     private Direction bedDir(Direction original, T entityLiving) {