+++ common/src/main/java/io/github/flemmli97/runecraftory/mixinhelper/MobNoAIHandler.java
@@ -1,0 +1,6 @@
+package io.github.flemmli97.runecraftory.mixinhelper;
+
+public interface MobNoAIHandler {
+
+    void setIgnoreNoAI();
+}